refactor: move downloader errs to own module
