+++ crates/interfaces/src/p2p/bodies/mod.rs
@@ -1,4 +1,8 @@
 /// Traits and types for block body clients.
 pub mod client;
+
 /// Block body downloaders.
 pub mod downloader;
+
+/// Error types.
+pub mod error;
